[refactor] rabbit mq config common 이관 (#22)
