+++ src/main/java/gdsc/cau/puangbe/common/config/RabbitMq/RabbitMqProperties.java
@@ -1,4 +1,4 @@
-package gdsc.cau.puangbe.photorequest.config;
+package gdsc.cau.puangbe.common.config.RabbitMq;
 
 
 import lombok.AllArgsConstructor;
